+++ codemods/deprecateUtilityComponents.js
@@ -50,7 +50,7 @@ module.exports = (file, api) => {
       attributes: {
         borderWidth: '1px',
         borderStyle: 'solid',
-        borderColor: 'border.primary',
+        borderColor: 'border.default',
         borderRadius: 2
       }
     }